+++ src/main/java/com/jeeeun/demo/repository/ProductOptionRepository.java
@@ -3,5 +3,5 @@ package com.jeeeun.demo.repository;
 import com.jeeeun.demo.domain.product.ProductOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductOptionRepository extends JpaRepository<ProductOption, Integer> {
+public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
 }
