+++ src/main/java/com/jeeeun/demo/repository/ProductVariantRepository.java
@@ -3,6 +3,6 @@ package com.jeeeun.demo.repository;
 import com.jeeeun.demo.domain.product.ProductVariant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductVariantRepository extends JpaRepository<ProductVariant, Integer> {
+public interface ProductVariantRepository extends JpaRepository<ProductVariant, Long> {
 
 }
