+++ src/main/java/com/jeeeun/demo/service/product/model/ProductResult.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 
 @Builder
 public record ProductResult(
-        Integer id,
-        Integer categoryId,
+        Long id,
+        Long categoryId,
         String name,
         String description,
         BigDecimal salePrice,
